chore: 인벤토리 추가, 가격, 수량 int 로 수정

+++ server/src/main/resources/db/migration/V5.0__order_ddl.sql
@@ -16,7 +16,7 @@ create table order_line_item
     id          bigint auto_increment primary key comment 'ID',
     order_id    bigint null comment 'order_id',
     item_id     bigint       not null comment '상품 ID',
-    order_count tinyint      not null comment '주문갯수',
+    order_count int          not null comment '주문갯수',
     name        varchar(255) not null comment '상품명',
     price       int(11) not null comment '상품 가격'
 );
